pyproject.toml
[tool.poetry]
name = "cgpos"
version = "0.9.0"
description = ""
authors = ["Tejomay Gadgil"]

[tool.poetry.dependencies]
python = "^3.9"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
numpy = "^1.25.2"
jupyter = "^1.0.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
greek-accentuation = {git = "https://github.com/jtauber/greek-accentuation.git"}
scikit-learn = "^1.3.0"
jupyter-dash = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
